Moved two spacecraft specific functions from Hohmann transfer env to spacecraft class
